Guard on the size of the constant value and then remove unnecessary cast.

+++ src/protocol/bigreq.rs
@@ -95,7 +95,7 @@ impl TryParse for EnableReply {
         let (sequence, remaining) = u16::try_parse(remaining)?;
         let (length, remaining) = u32::try_parse(remaining)?;
         let (maximum_request_length, remaining) = u32::try_parse(remaining)?;
-        if response_type as u32 != 1 {
+        if response_type != 1 {
             return Err(ParseError::ParseError);
         }
         let result = EnableReply { sequence, length, maximum_request_length };
